Set specver to produce desired rpm version
Since we earlier built an "8.ew" version of resource-agents that
returns an error if the slave IO thread isn't running, we now build
a new "9.ew" version that doesn't have that behaviour since it
caused problems during failover. While this version is identical to
the "v3.9.2" tag, we want a distinct (higher) version of the rpm
to simplify upgrades.